refactor(login.component.ts,session.service.ts): Moved login component and session service outside of the session module and changed login to standalone
