+++ src/BurstChat.Client/src/app/components/session/login/login.component.scss
@@ -1,8 +1,8 @@
-@import '../../../../../scss/theme';
-@import '../../../../../scss/loader';
+@import '../../../../scss/theme';
+@import '../../../../scss/loader';
 
 .login-card-container {
-    background: url(../../../../../assets/waves.svg);
+    background: url(../../../../assets/waves.svg);
     background-repeat: no-repeat;
     background-position: bottom;
     display: flex;
