+++ src/BurstChat.Client/src/app/modules/session/components/register/register.component.ts
@@ -4,7 +4,7 @@ import { faDragon } from '@fortawesome/free-solid-svg-icons';
 import { tryParseError } from 'src/app/models/errors/error';
 import { Registration } from 'src/app/models/user/registration';
 import { NotifyService } from 'src/app/services/notify/notify.service';
-import { SessionService } from 'src/app/modules/session/services/session-service/session.service';
+import { SessionService } from 'src/app/services/session/session.service';
 
 /**
  * This class represents an angular component that enables a user to register an account in BurstChat.
