+++ src/BurstChat.Client/src/app/modules/session/components/reset-password/reset-password.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 import { faDragon } from '@fortawesome/free-solid-svg-icons';
 import { tryParseError } from 'src/app/models/errors/error';
 import { NotifyService } from 'src/app/services/notify/notify.service';
-import { SessionService } from 'src/app/modules/session/services/session-service/session.service';
+import { SessionService } from 'src/app/services/session/session.service';
 
 /**
  * This class represents an angular component that helps the user to reset his password.
