+++ src/BurstChat.Client/src/app/services/session/session.service.spec.ts
